Add testbench header file
This C++ header is the bases for the pybind11 wrapper to a Verilator module.
Without alteration the header file is used in combination with a custom C++
source code file to provide a generic interface with the Verilator outputs and
produces a Shared Object that is acessessable from within a Python script.