Fill out axi4lite module
As with all the 'vpw' interface background objects the axi4lite 'Config' class
returns a Generator when the 'init' function is called by the background
registering function. Unlike the other background object its methods come in
two types. The more normal methods are non-blocking, designed to be called
between 'ticks' of the clock. The other methods, however, are BLOCKING and thus
can tack multiple ticks to complete.