Update tick in vpw.util to remove temporary tasks
So far, all generators in the 'vpm' package have been long lived and need to
exist for the length of the testbench. However, a testbench may want to have a
generator or composition of multiple generators that is only short lived. In
such a scenario creating a generator that is registered to the background but
will be removed once complete is ideal.