Move verilator header parser code into its own function
This separation of the parser and the create function allows for better testing
and debug of the auto generate interface code used by vpw. It also means that
the VPW package can be more easily integrated into a range of external build
tools instead of shelling out raw build commands like the create function does.