Add 'pause' streaming functionality to axis.Master
Like how the axis.Slave can backpressure the stream using the 'ready' function
to set the ready flag in the AXIS interface, this new 'pause' function can
deassert the valid flag of the master stream. However, the valid will only be
deasserted after a transmission thus making it easier to ensure the correct AXI
protocol.